+++ src/app/components/home/child components/chats/child components/chat/chat.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { ChatListItem } from '../../../../../../models/chat/chat.model';
 import { ChatApiService } from '../../../../../../api/chat/chatApi.service';
-import { RefreshDataService } from '../../../../../../services/refreshData.service';
+import { RefreshDataService } from '../../../../../../services/refresh/refreshData.service';
 
 @Component({
   selector: 'app-chat',
