+++ src/app/components/home/child components/contacts/contacts.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { ContactApiService } from '../../../../api/contact/contactApi.service';
-import { RefreshDataService } from '../../../../services/refreshData.service';
-import { ChatSocketService } from '../../../../services/chatSocket.service';
+import { RefreshDataService } from '../../../../services/refresh/refreshData.service';
+import { ChatSocketService } from '../../../../services/chat/chatSocket.service';
 import { ContactOnlineStatus } from '../../../../../../../common/dto/contact.dto';
 
 @Component({
