+++ src/app/components/home/child components/settings/settings.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { RefreshDataService } from '../../../../services/refreshData.service';
+import { RefreshDataService } from '../../../../services/refresh/refreshData.service';
 import { Router } from '@angular/router';
 import { UsersApiService } from '../../../../api/user/usersApi.service';
