+++ src/app/components/home/child components/show-chat/child components/messages/messages.component.ts
@@ -8,7 +8,7 @@ import {
 } from '@angular/core';
 import { Message } from '../../../../../../../../../common/dto/message.dto';
 import { MessageApiService } from '../../../../../../api/message/messageApi.service';
-import { ChatSocketService } from '../../../../../../services/chatSocket.service';
+import { ChatSocketService } from '../../../../../../services/chat/chatSocket.service';
 import { EVENTS } from '../../../../../../../../../common/constatns/gateway.contants';
 
 @Component({
