+++ src/app/components/login-page/login.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { LoginDto } from '../../../../../common/dto/user.dto';
 import { Router } from '@angular/router';
-import { RefreshDataService } from '../../services/refreshData.service';
+import { RefreshDataService } from '../../services/refresh/refreshData.service';
 import { UserCookieApiService } from '../../api/user/userCookieApi.service';
 import { UsersApiService } from '../../api/user/usersApi.service';
 // ...existing imports...
